requirements.txt
numpy
torch
pandas>2.2.2
pyarrow
duckdb==0.9.2
hydra-core==1.3.2
snakemake==8.16
tqdm
ipykernel
geopandas
scipy
